# backend/api/docs/swagger/swagger.yaml
openapi: 3.0.0
paths:
...
  /v1/tarot/cards:
    get:
      operationId: TarotController_getCards
      parameters:
        - name: type
          required: false
          in: query
          description: Tarot Type
          schema:
            enum:
              - MAJOR
              - MINOR
            type: string
        - name: suit
          required: false
          in: query
          description: Tarot Suit
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: Limit
          schema:
            type: number
      responses:
        "200":
          description: List of Tarot Cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TarotCardsResponseDto"
      tags: &a11
        - tarot
  /v1/tarot/cards/{id}:
    get:
      operationId: TarotController_getCard
      parameters:
        - name: id
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Get Tarot Card by ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TarotCardDto"
      tags: *a11
...
info:
  title: hieudoanm
  description: Hieu Doan
  version: "1.0"
  contact: {}
...
servers: []
components:
  schemas:
...
    TarotCardDto:
      type: object
      properties: {}
    TarotCardsResponseDto:
      type: object
      properties:
        total:
          type: number
          default: 0
        types:
          type: string
          enum:
            - MAJOR
            - MINOR
        suits:
          type: array
          items:
            type: string
        cards:
          default: []
          type: array
          items:
            $ref: "#/components/schemas/TarotCardDto"
      required:
        - total
        - types
        - suits
        - cards
...